Several changes related to cobbler import.
First, fixed an import bug in orphan detection.

Other features are related to running createrepo after a partial rsync to create nicer
repos, and shipping a more strict rsync.excludes file by default.  Note that the createrepo
change still needs to be applied to the "cobbler reposync" command, which works outside of import.