Further work on interchangeable backends.  Rather than keeping the config file format choice in settings (which is something of a Catch-22 situation), this may end up being a config file setting in /etc.
Module loaders work and still default to yaml, though I've coded up a sample
simple_json serializer that will be functional if users install simple-json.
This is just demoware, JSON isn't replacing yaml and is (at least in this case)
not as readable because of the way it escapes slashes.

This is primarily to enable future work to integrate with other config
file formats, such as possibly getting some of the system info from LDAP.
Possibly.