Added the ability to undefine a symbol that is defined in a parent object.
This means that if a kernel option is set for a profile, and you want it unset
in a child system, you can say --kopts='!foo' to undefine the foo.  Similarly you
can add symbols and undefine others at the same time... --kopts='bar !foo baz=3' and
so on.

This is relatively fringe usage but was previously not possible.