Adds support for the XMLRPC changes for Cobbler 1.5.
Some additional features added as a result is the ability to debug a
connection. If Base.debug is set to true then all communication between
the client and server is sent to stdout.

Also fixed the list_systems.rb example to show the image on which a
system is based if it's not based on a profile.

And some minor cleanup to those particular source modules.