cobbler4j: More fixes for <<inherit>> and blending.
Previous version had a problem where if you committed an object it would
stop inheriting and just set all the parents settings on itself. Not
optimal.

Now when we lookup an object that supports blending we also lookup a
blended data map. Whenever a getter is called we check if a string
<<inherit>> comes back, if so we go to the blended data map and return
the final value originating from the parent.

Blended data map is refreshed after a commit, just like the normal data
map.