Don't warn needlessly when repo rpm_list is empty
Since our rsync, RHN (and formerly APT) repository mirrors don't support
partial mirroring with --rpm-list, we warn when a selective list is provided
and ignored. Since it's easier (using the standard user interfaces) to empty a
mistakenly-set list than to reset it to "", let's not warn when the list is
empty.