modify keep_ssh_host_keys snippet to use old keys during OS installation
In new Fedora releases (13 and on) anaconda can be configured to allow
ssh connection into the system during OS installation.
It generates new ssh keys, but we want to be able to use the old keys.
The best solution would be for anaconda to check if the original keys
were saved before generating new ones. In the meantime we just wait for
the new set to be generated and replace them with the old set.
The patch is attached, it is also at https://github.com/lflaks/cobbler
--

Leon

Signed-off-by: Scott Henson <shenson@redhat.com>