Implement system power status API method and CLI command
Autotest (autotest.kernel.org) is beggining to leverage Cobbler to
auto install host systems. It happens to be valuable to our project
to be able to determine the power status of systems, so that we can
better determine if the systems are unresponsive or shut off.

This patch was tested using only APC power switches. Behavior and
command output is expected to be the same on other fencing agents.

The API call return values are probably the big disruption
point comparing it to the other power related calls. I find the
proposal (on this patch) to provide more clear and meaninful
information, but I do acknowledge it's disruptive. Feedback welcome.

Signed-off-by: Cleber Rosa <crosa@redhat.com>
Signed-off-by: Scott Henson <shenson@redhat.com>