Create more condensed s390 parm files
This change allows the utilization of all 80 characters per line in
s390 parm files which makes them more condensed.

It should now also be possible to use parameters that are longer than
80 characters like for example long installation urls and similar.

I have removed the strip function because it is not needed anymore and
might also result in problems with this change.
The leading space could be also removed with an additional if else
expression but I think it is not worth it.

Signed-off-by: Tim Hardeck <thardeck@suse.de>
Signed-off-by: Scott Henson <shenson@redhat.com>