Hello,
I tried to kickstart a SLES 11 distribution on spacewalk, but this was
not possible for me.

In spacewalk the kickstart profile have a special name: <distribution
name>:<id>:<organisation name>. The problem is the ':' in the name.

When you kickstart a SLES 11 distribution it will url encode the URL,
and change the ':' into '%3a'.

Because the services.py script does not url decode the URL, this will
not work and you get the error '# profile not found'.

I added the urldecode to the services.py script an attached the patch.

Kind regards,

Jasper Aikema