Tim Verhoeven <tim.verhoeven.be@gmail.com> (Tue. 08:35) (Cobbler attachment) Subject: [PATCH] Add support to koan to select type of network device to emulate To: cobbler development list <cobbler-devel@lists.fedorahosted.org> Date: Tue, 2 Aug 2011 14:35:21 +0200
Hi,

Attached is a small patch to koan that allows you to select what type
of network device kvm/qemu will emulate when creating the guest. It is
based on the --qemu-disk-type option, but it is called
--qemu-net-type. I need it to be able to select the virtio type, but I
listed the other possibilities in the help output.

Now it would be great if this could also be added to Cobbler itself so
it can be stored as part of the profile and/or system definition. But
that it a bit more work I guess.

Regards,
Tim

Conflicts:

	koan/app.py