bind zone template writing
Hello,

I have downloaded Cobbler 2.0.11, and have found the following bug. In
the BIND zones being written by Cobbler the RR lines are not
terminated with a semicolon. ISC BIND likes to have (ie requires) the
semicolons there. Here is my patch to correct the problem. It works
well to fix the problem in my setup.

-jessec

Signed-off-by: Scott Henson <shenson@redhat.com>