Fix puppet certificate names when hostname != fqdn
Puppet certificates are named after the fqdn, not the hostname. When run on a
host where the hostname doesn't look like an fqdn, Puppet (actually Facter)
assumes that that is the short hostname, and uses the first domain in the
resolver's search list as the domain name to make the fqdn. (cf.
facter/host.rb, facter/domain.rb, facter/fqdn.rb). The cobbler module should
do likewise.