koan: port imagecreate to virt-install
This code was using the API equivalent of the virt-image tool,
however it's really unnecessary here, all we are doing is importing
an existing disk image which is an install option long provided by
virt-install. Merge the functionality into the virtinstall module
and convert it.

This code has actually been broken for about 6 months, since the
start_install signature was never updated for the qemu_net_type
change.

Signed-off-by: Cole Robinson <crobinso@redhat.com>