koan: Port xen creation to virt-install
We break out the virt-install logic to a separate file, since further
conversions are going to be able to share 90% of the code (basically
the whole point of libvirt :) ).

In the process we drop a lot of hacks and workarounds for things that
have been fixed in virtinst for at least a year, some for as long as
3. The required virtinst dep is fairly recent anyways with the
call to guest.set_autostart.

v2:
    Return command in format expected of subprocess.call
    Use proper shell splitting for extra args

Signed-off-by: Cole Robinson <crobinso@redhat.com>