fix infinite netboot cycle with ppc64 systems
I encountered a rather annoying problem with cobbler for ppc64 systems:
after an install succeeded with pxe_just_once set, the DHCP
configuration file/DHCP server process were left alone. The DHCP server
would then respond on the network on the next boot and the target server
never booted from the disk. The netboot is controlled by SMS/yaboot and
if SMS receives a DHCP response for a file to tftpboot (in this case the
file being yaboot), it will stay in an infinite loop within the netboot
code, even if the netboot eventually fails. That code exists in the
field as-is, so fixing it, while worthwhile (if possible), doesn't help
with existing deployments.

I worked around this by only generating the DHCP configuration if a
system does have netboot_enabled set and re-generating the DHCP files
via a new API, sync_dhcp() whenever disable_netboot() is called.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>