snippets: save/restore boot-device on ppc64 on fedora17
I previously added some example snippets for sles11 to support saving
and restoring the boot-device on ppc64, as the sles11 installer
overwrites this information. The fedora17 installer also does this now,
so the snippets need some extension, as the installer disk layout is
slightly different. Handle the two known cases distinctly.

Tested with SLES11-SP2 and FC17-prebeta3 on ppc64.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>