distro_signatures: add SLES11 ppc64 support
SLES has a slightly different signature from OpenSUSE. I only have
access to ppc64 for testing, so only indicate they are supported. Should
be easy to verify/test on other architectures and extend it later.
Similar to the OpenSUSE signatures, I've added the Service Packs of SLES
as distinct sub-breeds.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>