Rework the glob() mechanism.
Unfortunately this does not work anymore if we configure some files. The
problem is that these globs are evaluated when setup() is called but the
files are generated when then configure task is run.

So on the first run the configured files where not found and because of
that not installed.