Delete debuginator script
This script has a very specific purpose of automatically creating empty initrd/kernel
files for all existing distros, but script is broken (it reads distros in old YAML format
instead of newer JSON one).

Test scripts already create fake initrd/kernel files when needed. I do not see other
use cases of this script. The fact that it is broken for a long time is one signal
that this is not needed anymore.