Set ksdevice kernel option to MAC address for ppc systems
Setting the kernel option "ksdevice" to "bootif" tells the Linux installer
to use the MAC address set in the "BOOTIF" kernel option. It is not useful
for ppc systems, since, unlike pxelinux, yaboot does not add the "BOOTIF"
parameter automatically.

As it is not used, it may occur that the installer prompts for the interface
to be used (for RHEL 6.x distros and with more than one interface).
Such scenario prevents a fully automated provisioning.

Signed-off-by:
 - Alan Evangelista <alanoe@linux.vnet.ibm.com>
 - Andrea Bucci <andreavb@linux.vnet.ibm.com>
 - Eduardo Bacchi <ebacchi@linux.vnet.ibm.com>
 - Nathan Ozelim <natoze@linux.vnet.ibm.com>