Prevent disappearing profiles after cobblerd restart (#1030).
This diff makes the use of --parent and --distro mutually exclusive.
The consequence is that subprofiles always have the same distro as
the parent profile. I think this has been the intended behaviour
ever since subprofiles got introduced, now it's just better enforced.

This diff also checks for assciated subprofiles first when you try to
delete a parent profile, this should prevent data corruption. Of
course using the 'recursive' flag is still supported.

Cobbler-web has been adjusted to use the (safer) xapi_object_edit API
for the deletion of items.