gpgcheck / enabled are not valid in kickstart, only in config.repo
Commit 6f2b3f5b93ec7681f730dc6b847f7bac6decc038 changed kickstart
generation to add all defined yum options for a repo to the
repo line in the kickstart file. However, only some options are
valid in the kickstart (and this no doubt changes with anaconda version).
In particular, gpgcheck is allowed in the yum configuration file, but
not on a kickstart repo line.

Allow gpgcheck to be overridden from the cobbler default,
forcing gpg signature verfication in accordance with STIG requirements,
without including gpgcheck in the kickstart.

Also, never include "enabled" on the repo line.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>