Remove methods which create items from Config class
To keep code simpler, item creation should always be done
using the same method. Currently, there are 2 ways to create
an item: creating it directly or using Config class. The
purpose of this patch is to eliminate one of those and make
item creation more consistent.

There is no gain in keeping item creation in config class,
as it only creates the item, it does not do anything
additional such as adding the item to its respective collection.