Improve kssendmac kernel option handling
Before, in blender() method, if distro breed was redhat, kssendmac
kernel option was added if kernel command line was smaller than 244
chars. First, this should be handled in tftpgen.py, not blender()
method. Second, this size test is not good because it considers the
command line without the kernel options which are automatically added
by tftpgen.py. Third, if kssendmac kernel option is optional and automatic
installation would still work without it, it should not be added
automatically by Cobbler. I suppose it is a required option. For last,
Cobbler already outputs a warning if kernel command line smaller
maximum size (255 chars) is surpassed.

New implementation always adds kssendmac kernel option in tftpgen.py
if breed is redhat.