Add script to create a KVM guest
Also, the script installs a Red Hat based Linux distribution
manually or in automated fashion using a kickstart.

The script will be used for creating a Cobbler test environment
in automated fashion.

Also, add a simple Cent OS 6.5 kickstart so it can be used to
easily create a test environment. url and possibly network kickstart
directives will be replaced when script which creates a KVM guest
is run.