distro_signatures: rhel7.1 adds ppc64le as a supported architecture
RHEL7.1 adds support for little endian ppc64. Simply add it to the valid
list of architectures for the rhel7 signature. It won't get picked up on
import of rhel7.0, as only ppc64 is present on the RHEL7 ppc64 ISO.
There are separate ISO images for RHEL7.1 (ppc64 and ppc64le) and need
to be imported as distinct distros/profiles.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>