manage_import_signatures: search for architectures in reverse length order
For RHEL7 on ppc64, for example, Cobbler ends up creating both a ppc and ppc64
entry. Leaving aside the fact that I don't think 32-bit ppc is supported by
RHEL7 (at least not with this iso, that's for sure), the issue is that both
"ppc" and "ppc64" (both listed in signature["supported_arches"]) are "found"
within the signature["kernel_arch"] file which contains "ppc64".

Two fixes are needed:

1) self.get_valid_arches() should return a reverse-sorted list by length
(longest first)
2) If a match is found (foo[arch] will be set to 1), the arch loop should be
broken out of, as the first match found is the most specific (matches the most
characters)

Fixes  https://github.com/cobbler/cobbler/issues/1332

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>