cobbler: add powerpc grub2 support
GRUB2 is a highly modular and flexible replacement for GRUB. Newer
distributions (FC20+, RHEL7+, SLES12+, Ubuntu 14.04+) support GRUB2 on
Power, and yaboot is now deprecated there. Add "grub2" as a valid
boot_loader type, and provide a template for its config file. GRUB2
will only be used if a signature on import indicates the distribution
should use it instead of yaboot. Eventually, we would probably remove
yaboot support (or switch the default around, at least).

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>
---
Jorgen, I also need to provide you with a binary grub2 tarball that can
be used by get_loaders, I think. What is the process to do that, as they
are not in the repository?