cobbler: add powerpc pxelinux support
In PowerNV (non-virtualized) mode, certain Power8 systems run petitboot
which can use pxelinux for network booting. Add support throughout the
code for pxelinux config file generation for powerpc systems. There is
one distinction between pxelinux on Power and on other systems, though,
the bootloader is built into the firmware, so we do not want to send any
file via the DHCP configuration.

There is currently a bug in petitboot that misparses 'ipappend 2' and
doesn't prefix the machine type, e.g. "01-". I'm manually working around
this in tftpgen, and will undo that change once petitboot is fixed.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>