Add boot loader field to distro
Allow user to override which boot loader he/she wants to use to netboot
a distribution instead of always automatically selecting it. TFTP/DHCP
files generation can now know which boot loader was selected.

It is needed for distros, because some distros, particularly
little-endian powerpc ones, need to use grub2, while historical
distributions can continue to use yaboot. Additionally, PowerKVM
installations need to be able to use pxelinux.

Note that I'm currently working around an issue I hit with the presence
(or lack) of utils.SIGNATURE_CACHE. In manage_import_signatures.py,
utils.SIGNATURE_CACHE is {}, because (I think) it's loaded before the
signatures are in CobblerAPI. If we were to fix that issue, we could
simplify the utils code, which currently handles two call-paths, one of
which has the SIGNATURE_CACHE available and one that doesn't (thus
relying on the API function to get to it).

Many thanks to Alan Evangelista <alanoe@linux.vnet.ibm.com> for comments
and review on this patch as well as the original functionality to add
boot loaders to system objects.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>