Add boot loader to system
Allow user to override which boot loader he/she wants to use to netboot
a system instead of always automatically selecting it. TFTP/DHCP file
generation code can now know which boot loader was selected now that the
boot loader selection can be overridden.

It is needed for systems, because on certain Power8 systems running in
baremetal, petitboot is used to netboot and requires a pxelinux
configuration regardless of distribution (which might use yaboot or
grub2 normally). The platform ("PowerNV") could be used to determine
this, but currently Cobbler doesn not have a notion of platform.

Based off a patch by Alan Evangelista <alanoe@linux.vnet.ibm.com>.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>