Ubuntu: save/restore boot order on ppc64el
Ubuntu is like RH and SUSE in that it overwrites the boot order upon
installation. This affects automated testing as we can't simply power
cycle the machine to netboot as expected. Add support to the snippets
for saving/restoring the boot the boot device list via the nvram
command.  Additionally, add a non-chrooted script to autoinstall_scripts
and run that in the target system, but outside of /target. This is
necessary so we have access to the same data as in the early script.

Signed-off-by: Tomo Berry <tbberry@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>