system: allow systems to override the bootloader to grub2
We currently can override a distro's bootloader to use pxelinux on a
per-system basis. However, if you specify any other bootloader (e.g.,
grub2), yaboot is used instead. For certain environments, it is useful
to use grub2 rather than yaboot, even if yaboot is specified as the
distro-supported bootloader. This can happen, for instance, to work
around firewalls, as grub2 supports more network options (http, etc.).
Allow a system to override the bootloader to grub2, rather than only to
pxelinux.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>