cobbler: fix get-autoinstall cli interface
Because the CLI is proxied via the RPC interface, we cannot use named
parameters to the remote APIs. We have to either pass empty strings in
the right position, which is both fraught with likelihood of breakage,
but also breaks expectations about the default values (e.g., None), or
introduce helper APIs that are clear and simple. In this case, add a
simple helper interface that calls generate_autoinstall() with the named
parameter.

Fixes: 7ef12d1b ('Create automatic installation files manager')
Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>

---

Note that if the mentioned commit simply did one thing, instead of many
things, there would not have been a regression. E.g., if it had just
reorganized the code instead of changing any callers of APIs.