Fix RHEL6 and RHEL7 bonding configuration
RHEL6 and RHEL7 don't have an /etc/modprobe.conf, so the bonding options
never get specified. Instead, they either use
/etc/modprobe.d/bonding.conf or automatically start bonding based on
options within the master bond interface.

This patch does the following:

If we're on RHEL6 or 7, we use /etc/modprobe.d/bonding.conf, otherwise,
/etc/modprobe.conf

If we specify bonding opts, we don't actually write to the modprobe
config, as RHEL is intelligent enough to start the insmod the bonding
module if it detects it. Otherwise, you need the 'alias' commands. This
works for RHEL5, 6 and 7.

For RHEL7, the default is to use NetworkManager, so we have to specify
the settings TYPE=Bond and BONDING_MASTER=yes on the bond interface
config file.