Add "airgapped" option for standalone ISO to also include repos
This commit incorporates the method described in this document:

https://github.com/cobbler/cobbler/wiki/Adding-Extra-Repositories-to-%22buildiso-standalone%22-Media

The issue, as described there:
-----
The "cobbler buildiso --standalone" command will create an ISO
image of a DVD which can be used to install the selected
distribution. However, it does not include any additional
repositories which you might have defined in the profiles - only
the base distribution. This means that if you have an errata
repository or additional packages from something like Red Hat MRG,
those RPMs won't make it into your installed system.
-----

The "repo" directives, expanded from the $yum_repo_stanza keyword
in the template by the generate_repo_stanza() function in the
autoinstallgen.py file, use an HTTP link directed to the
"http_server" value for the repo. This URL will be unreachable
on an install target which is not connected to the network or
cannot reach the specified server, causing the installation
to fail when a package from one of those repos is requested
during the installation. This limits the usefulness of the
standalone ISO, since if you can reach the repos you can reach
the distro too and might as well do a non-standalone buildiso.

With the "cobbler buildiso --airgapped --distro=XYZ," the
standalone ISO will be constructed to include all repositories
referenced by any system or profile within the XYZ distro. This
will allow installation on an "airgapped" system completely
disconnected from the network on which the Cobbler server is
located.

If a repo is not configured for a local mirror, the operation
will error out - the repo contents are copied using the
utils.rsync function. It would be possible to invoke a reposync
directed to the buildiso directory for non-local repos, but
the amount of time needed and the lengthened list of failure
modes were a consideration in imposing this local-mirror-only
requirement.

The repos are copied into the repo_mirror directory in the
ISO build directory, which differs from the original instructions
which had you place the repos in "cobbler/repo_mirror" on the
media. The cache for the repos is excluded, because the cache file
names have a tendancy to overflow the Joliet file name conversion
tables as described in the document.

The code takes no notice of the size of the repos. It will happily
build a 20GB ISO image if you include enough repos in your profile,
trusting that you either have a BluRay burner (and readers on your
target systems), or can use virtual-DVD capabilities of
enterprise-class machines. It will also happily fill up your
filesystem if you don't specify paths with sufficient disk space.

No provision is made for cobbler_web use of this capability,
as it is basically just an extension of the --standalone option
which is unsupported in the web interface. A web form for building
ISO images would be a useful feature.