Refactor profile and system filtering in generate_netboot_iso
When the user specifies a list of profiles and/or systems for an
ISO image, the full list of profiles and systems is filtered down
to include only those specified.

This commit creates a new function, filter_profiles_or_systems(),
which when given a list of names and the type - either profile or
system - will return a list of profile or system objects
containing only those which were both specified and existing.

If no filter list is provided, the function returns all objects,
as before. If none of the names provided match any valid profiles
or systems, the function will utils.die() rather than producing
a useless isolinux.cfg.