Include only selected profiles in standalone ISO build.
Function generate_standalone_iso now recognizes the --profiles
argument, which limits the profiles and systems included when
building the ISO image.

This allows a user to build an ISO which includes only the
desired profiles, and prevents inclusion on an airgapped ISO
image of repos which might be referenced in other profiles
under the same distro but are not referenced in the selected
profiles.