Added support for fixed DHCP IPs when using vlan over bond.
In my server config I use em1+em2 -> bond0 -> bond0.XXX vlan setup.
Additional bond0.YYY are created later on with puppet. Them main
problem is that some vlans have MTU 9000 and therefore they need bond0
(and em1 and em2) to have MTU 9000 as well. The system/installation
vlan XXX is, however, with MTU 1500 only so I cannot put the IP on the
bond0 interface directly.

So far cobbler did not create fixed address entries in DHCP
("fixed-address X.X.X.X;") if there was no IP on emX or bond0 but
only on bond0.XXX. This fix adds the fixed address from the 'first'
vlan interface on top of the bond device, if the bond device itself
does not have an IP.