Add optional serial configuration to grub and pxe files
Systems can now have an optional serial device number and/or serial baud
rate. If one of the two values are given, the other will be set to a
default value (unit: 0, baud rate: 115200). If so, a leading 'serial'
line gets added to grub/pxe files for the specific system.

If none of the values is set, nothing changes.