Cleanup grub2 vs grub and provide new grub*.template files
There is no grub2 string anymore anywhere in the project
(maybe in a comment...).
grub means grub version 2 aka grub2. We use consistent "grub" naming
for this from now on.

/var/lib/cobbler/grub_config has been introduced (from config/grub/*)
which holds the base tree of static grub configuration files and
this directory is (r)synced the same way to /srv/tftp as the loaders
dir holding pre-compiled boot loaders and grub modules.