scm_tracker shouldn't read additional setting when disabled
Seems like it might happen that we don't have a setting for push_script. I
guess this would be fine, if scm_tracker is disabled anyway. But I don't
know the the settings needs to be there otherwise.

With moving the check for scm_tracker_enabled above reading the rest of the
settings, exceptions can be prevented when a settings for scm_push_script
is not there.