Introduce remote-boot-{kernel,initrd} to be passed as parameter to a distro
One can then pass a URL pointing to the installation kernel/initrd of a
distribution. If the bootloader has this support, it will directly download
the kernel from this URL, instead of the directory of the tftp client.

Note: The kernel (or initrd below) will still be copied into the image
directory of the tftp server.
The --kernel/--initrd parameter is still needed (e.g. to build iso images).
The advantage of letting the boot loader retrieve the kernel/initrd directly
is the support of changing/updated distibutions. E.g. openSUSE Tumbleweed is
updated on the fly and if cobbler would copy/cache the kernel/initrd in the
tftp directory, you would get a "kernel does not match distribution" (or
similar) error when trying to install.