Pass distro variables to setup.py via environment variables
I could not find a way to use global options like:
python3 setup.py --my_option
setuptools seem not to be made for this.
But this is needed because e.g. webroot is directly passed to setup(..)
and at this point of time, no arguments parsing happened yet.

I could split out httpd_user as install options, but this would have made
this patch much bigger and complicated.