Pass distro variables to setup.py via environment variables
This patch consolidates distro specifics.
Currently there are different:
- Package names
- Apaches user/group
- Lots of different file system paths

Before this info was spread across:
- setup.py
- Makefile
- spec file

Now this info is laid out and centralized into:
distro_build_configs.sh

For the spec file this unfortunately only works
for variables needed in the %build or %install section.

apache user and group for example is needed in the %file
section and can therefore not be sourced from outside.

But it is now double checked to be consistent. An error
is thrown in the %build section if data which needs to be
defined in the spec file and in distro_build_configs.sh for
setup.py or Makefile does not match.