Update validate.py
Copy function in collection.py will set the mac address on a system collection to a blank string ("") this is then validated. Previous versions of cobler had specifically accepted "" as a valid mac.

Attempting to copy in newer cobbler is throwing an exception that the blank mac address generated is invalid