Better performing string replacements
Previously we where iterating over the whole search table to do string
replacements. This can and will in most of the cases be very wasteful.

In order to prevent this, a regex lookup is now performed that returns
the keys that are used in the template and only for those keys we do a
str.replace().