feat: Add option to sync only specified systems
If managing a vast amount of systems via a single Cobbler instance  the
amount of time a `cobbler sync` call needs to synchronize everything is
quite a lot. @cxfcxf observed that writing PXE configuration files for
every system to the tftpboot location takes up to 90% of the time of a
`cobbler sync`. So he introduced the option to synchronize only
specified systems with `cobbler sync --systems=system1,system2`.

This also adds some basic unit tests and unit test skeletons which has
to be completet in the future.