fix: Use of 'next_serverv6' and 'next_serverv4'
This will split the 'set_next_server' method in system.py and
profile.py into 2 separate methods for IPv4 and IPv6. Furthermore it
introduces the 'VALUE_INHERITED' constant in item.py to unify the use
of '<<inherit>>'. At the moment this constant is only used in the new
next server methods.