Do not check if a file exists before removing
Apart from TOCTOU races this has another implication: if a symlink is
pointing to a non-existent file, os.path.exists() will return False,
hence the link will never be removed. In the tftpgen case this will lead
us to a deadlock if we have inconsistent Grub configuration: it's not
possible to create a new link before removing the old one, and we cannot
remove the old one because the existing link is broken.

Both utils.rmfile() and utils.mkdir() can handle ENOENT and EEXIST
respectively.